backend/requirements.txt
flask==3.1.0
flask-cors
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
